server/app/requirements.txt
aiokafka==0.12.0
aiosqlite==0.21.0
amqp==5.3.1
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
billiard==4.2.2
blinker==1.7.0
celery==5.5.3
certifi==2025.7.9
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.4
dotenv==0.9.9
fastapi==0.120.0
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
future==1.0.0
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
idna==3.11
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.5.4
MarkupSafe==2.1.5
matplotlib==3.9.0
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.0
pgvector==0.4.1
pillow==10.3.0
platformdirs==4.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
pygame==2.5.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytube==15.0.0
redis==7.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
sqlparse==0.5.0
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.14
Werkzeug==3.0.1
